wire: move towire/fromwire_short_channel_id out to bitcoin/short_channel_id.c
