+++ gossipd/gossip_peerd_wire.csv
@@ -1,4 +1,6 @@
 # These must be distinct from WIRE_CHANNEL_ANNOUNCEMENT etc. gossip msgs!
+#include <bitcoin/short_channel_id.h>
+
 # Channel daemon can ask for updates for a specific channel, for sending
 # errors.  
 msgtype,gossipd_get_update,3501
