+++ wire/towire.c
@@ -146,19 +146,6 @@ void towire_secp256k1_ecdsa_recoverable_signature(u8 **pptr,
 	towire_u8(pptr, recid);
 }
 
-void towire_short_channel_id(u8 **pptr,
-			     const struct short_channel_id *short_channel_id)
-{
-	towire_u64(pptr, short_channel_id->u64);
-}
-
-void towire_short_channel_id_dir(u8 **pptr,
-				 const struct short_channel_id_dir *scidd)
-{
-	towire_short_channel_id(pptr, &scidd->scid);
-	towire_bool(pptr, scidd->dir);
-}
-
 void towire_sha256(u8 **pptr, const struct sha256 *sha256)
 {
 	towire(pptr, sha256, sizeof(*sha256));
